Add platform config for tqtc-installer-framework
This is similar to installer-framework.yaml with added configurations
for QtInstaller specific builds.

This enables two sets of builds from the same branch of
tqtc-installer-framework.

First set is for plain tqtc-installer-framework builds
i.e. the commercial deliverable.

The "QtInstaller" set is for the consumption of Qt Installer
creation.

Task-number: QTIFW-1027
Change-Id: Ifff750eef2631ce4e16edb32a7f7df593c698944
Reviewed-by: Simo Fält <simo.falt@qt.io>