CMake: Update instruction file to accept cross-compiling configs
This is needed for consistency.

Previously if you scheduled a test build for a qt5.git change, the
cmake cross compiling configurations would not appear and not be
tested.

But when staging the change, each repo WOULD start building the
cross-compiling configurations, except for the qt5.git top-level
builds.

The current change will make sure that configurations is done both
during scheduled test builds and actual staged builds.

Change-Id: Ia2cbb08cee97414c39287a2ab9ff8d312dcd813a
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Liang Qi <liang.qi@qt.io>