Set QT_QUICK_BACKEND=software in QEMU configs
In Qt 6 things get more complicated because we have Vulkan, Metal and
D3D in addition to OpenGL. Qt Quick will only default to the software
backend when no graphics API was available at build time - and this
not true for the QEMU configs because the Ubuntu env has both OpenGL
and Vulkan headers, so they both are enabled, but are not usable at run
time. To simplify life, just state the Qt Quick backend we want the
tests run with.

Change-Id: I63cd9534bc8cd5d0568235bc2035f5613d29ced4
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>