CMake: Add some missing Linux host Android target architectures
In Qt 5 times, qmake was able to build libraries for multiple Android
ABIs in one build tree.

In Qt 6.0 + CMake we can't do that yet, which is tracked in
QTBUG-80943.

We still need to build and ship libraries for all the ABIs, so use the
old style approach of one configuration per ABI.

Add some of the missing ABIs for the Linux host case:
arm64-v8a, armeabi-v7a, x86.

The x86_64 ABI will be added in a separate change, due to a bug in
qtbase regarding Gui's qdrawhelper.cpp and avx2.

Because it's a lot of copy pasting, move the configs into a separate
yaml file so it's easier to search and replace in the future.

Task-number: QTBUG-86053
Change-Id: If8419ce20ac033b644c49c0bac70196cd30fd679
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Toni Saario <toni.saario@qt.io>