CMake: Remove CI qmake configs for Qt 6.1
Qt 6.0 is now built with CMake successfully in the CI and we have
most of the coverage we had with qmake CI configurations. Apart
from platforms and modules we deemed not part of 6.0 release.

We didn't have enough time to disable qmake configurations for
6.0.0, but we should be able do it now for 6.1.

We can also consider removing qmake configurations for 6.0.x
sometime later.

Remove all CI configurations that use qmake to build Qt.

Keep the default.yaml license check UseLegacyInstructions
configuration, because license checking hasn't been ported to use new
Coin instructions.

One consequence of this is that developers that are working on
platforms and modules not yet ported to CMake (like WebAssembly or
qtwebengine), will no longer have coverage in CI until they are ported
to CMake.

If there are repo owners that wish to retain CI coverage for building
their repo with qmake (until a CMake port is done), they can do so by
creating a product_dependencies.yaml file in the root of their repo
and specifying an older qt5.git sha1 (so before this change is merged)
which still contains qmake CI configurations.

Task-number: QTBUG-88741
Task-number: QTBUG-86053
Change-Id: I985cf0597d6d3e1717fdff8b654397a5f9236a6e
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>
Reviewed-by: Simo Fält <simo.falt@qt.io>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>