CMake: Update provisioned CMake to version 3.18.3
This version should fix the iOS PCH builds and Ninja Multi-Config,
as well as the issues with AUTOMOC and Ninja.

Temporarily explicitly disable the pkg_config feature on Windows
hosts, because otherwise the configuration will break. The
FindPkgConfig module in 3.18 has some additional behavior for checking
if the found pkg-config works, and that's not handled by our
qt_build_internals_disable_pkg_config_if_needed function in qtbase.

Fixes: QTBUG-86862
Change-Id: I80514d3d7896ae09ab252aed10e86725fae0cc9c
Reviewed-by: Toni Saario <toni.saario@qt.io>