CMake: Build iOS in a release and debug configuration
To ensure that examples can be built with an xcode project
generated by qmake, we have to ship both debug and release
libraries.

So use Ninja Muli-Config to build both configs.

Note we build Release and not RelWithDebInfo on purpose, so we
don't inflate the package size too much due to the embedded debug
information.

We can probably switch to RelWithDebInfo once we implement the
separate_debug_info feature for static builds.

Fixes: QTBUG-87638
Change-Id: I3e26cd0e47fc96fbdee8b3b0177f78fa266b8cc4
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>