Add worktrees as detached
Since we later check the worktree out to the revision specified
in the dependency file, add it as detached to avoid conflicts
with already existing worktrees for the default branch.

As a drive-bye, fix a typo in a comment.

Change-Id: I81c30c1fa3cfc9831fe7884a103746b43478dd8e
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
(cherry picked from commit 5bbec95c482e1afa9bf217f28f9bfc47a170b745)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>