Checkout as detached if checking out branch fails
Having several toplevel worktrees for the same branch is problematic, but
but we do want to be able to have one toplevel worktree synchronized to
e.g. one submodule's dev and another with all submodules at the respective
dev heads.

So fall back to detached checkout if regular checkout fails, and generate
a warning.

Change-Id: I13981ab931cbbd8b56187979708bfe63f7373718
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>