Provisioning: Update Squish for Qt6
These are snapshot packages from Froglogic, not available for macOS yet,
thus it will still use Qt5.15 based Squish. Only Linux and Win packages
are updated now.
Along with the Squish for Qt6 we need the Squish for Qt5 to be able to
run Creator tests (since Creator is still Qt5 based).

Change-Id: Iec0294e1de036db004633ee6b95e1e68a6dd23a5
Pick-to: 6.0
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Heikki Halmet <heikki.halmet@qt.io>