Change GCC version to 8.3.0 for Ubuntu 18.04
We need to use gcc 8 instead of 9 so that the Qt installer
will work with CentOS 8.3 which have gcc 8 in the system.
If we use newer gcc version with the binaries there will be issues
with libstdc++.
Using gcc-toolset-9 in CentOS won't help either because it will link
libstdc++ against the one in system.

Pick-to: 6.1
Task-number: QTBUG-91352
Change-Id: I2a3d186bc04b1f54f14ed95f0588d18bd63814c4
Reviewed-by: Ville Voutilainen <ville.voutilainen@qt.io>