Add Packaging target for openSUSE 15.2
With current approach we need to make two different linux installers.
One for SUSE and one for other linux targets we support. This is because
current CentOS version have newer GLIBC version than what we have in SUSE.
This means that installer made with CentOS won't work in SUSE versions
we support

Task-number: QTBUG-91873
Change-Id: If65d66093bf8dcb5fe46812e06d6b9194fdedac2
Reviewed-by: Jani Heikkinen <jani.heikkinen@qt.io>