macOS: Don't specify OpenSSL path for macOS configurations
Currently the qtbase configure system always chooses to use
SecureTransport on macOS configurations, even if a usable OpenSSL is
found.
As it is, there's no point to pass the OpenSSL location as a
configure argument because one might get confused to think that Qt
will be built with OpenSSL support (it won't).

There is work in progress to allow building Qt Network with support
for multiple ssl backends, but until that work is finished, remove the
configure argument to avoid confusion.

Change-Id: Ib91f0b85edc580145a78ab38aba7718e90a95240
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>
Reviewed-by: Tony Sarajärvi <tony.sarajarvi@qt.io>