CMake: Remove CMake minimum version check in top-level project
The check is done in qtbase instead.

Trying to do it in the top-level project is cumbersome due to needing
to know whether it will be a static or shared Qt build.

It's also cumbersome because the top-level build used to use qtbase's
minimum cmake version values and that can get awkward to handle with
submodule dependency updates.

Replace the check with a regular minimum required call.
The minimum version is 3.16 and the upper range value for NEW policy
assignments is 3.20, while 3.21.0 is still kind of fresh and has some
AUTOUIC issues.

As a drive-by, add a clarifying comment about QtAutoDetect.cmake.

Task-number: QTBUG-95018
Change-Id: Ifb4ec0b4ed7f56ab0179feb79072c93d1d66ce63
Reviewed-by: Craig Scott <craig.scott@qt.io>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
(cherry picked from commit 4180e9c5111faae9b242253e2ab215154174507d)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>