CI: Turn dbus support off for QNX_710
QtDBus is built even if system dbus is not found, and QtGui depends
on QtDBus.  To link a consumer of QtGui, the -rpath-link option must
be passed to the linker.  That is not done at the moment, due to
QTBUG-86533.

Fixes: QTBUG-96791
Task-number: QTBUG-86533
Change-Id: I381747f15135fe65777985742dca0fd102583941
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
(cherry picked from commit 420adf322a71aee5035cb6ae075e11afcee1fe3c)