Add script to create shared folders for testing Qt I/O with UNC paths
Once these paths exist reliably on the Windows VM running the test, and
are shared, we can access them through UNC paths as

\\${COMPUTERNAME}\testshare

This allows us to remove dependencies to network infrastructure from
those QtCore tests.

The test needs to run last, as otherwise the Visual Studio installer
mysteriously fails to execute. So make some space at the end of the
script list.

Pick-to: 6.4
Change-Id: Ie11b894f000722a7d99b2c5e279c1b0b63021003
Reviewed-by: Simo Fält <simo.falt@qt.io>