qt5: update qtlocation sha
qtlocation is an ignored module, because it's not yet officially
supported in Qt 6. However, it is checked out when someone clones the
repo and wants to build Qt from sources.
The sha of the checked-out commit for qtlocation is very old, the repo
contains both qtlocation and qtpositioning sources at that revision.
It is causing configuration errors, because we now have qtpositioning in
its own repo, so we get duplicated targets.

This commit bumps the revision of qtlocation submodule to the current
dev HEAD. This will eliminate the configuration problems.

Change-Id: Ia5de4d1357ea657b42465e473d5b6e02a2bb6f30
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>