Fix calling of script creating shared folders
The calling script for each platform missed the leading ". " so
powershell never executed the script. Sadly, provisioning still
succeeded.

Add some log output to the script doing the work to verify that
the file system is what is expected to be at the end. If it's not,
then the cd'ing into the shares will fail and the script will
return with a non-zero exit.

Pick-to: 6.4
Change-Id: I8fe85c7e202132525876cdddf41750732f69aaf9
Reviewed-by: Simo Fält <simo.falt@qt.io>