clang-format: codify some of our conventions
1. We tend to use 2 spaces for indenting pre-processor directives,
when we do it. Unfortunately clang-format also counts header-guards
so not really recommended to use blindly on a header...

2. Add comments to namespaces, since the closing brace is simply
a brace with no further decoration.

3. Attach escaped new-lines on the left (instead of current: right).
Where 'right' will add the '\' characters on the far-right of our column
limit, 'left' will put them as far left as it can while having them
vertically aligned.

Change-Id: Ib5a836c76f7761d743e0100c83e7db609b959249
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
Reviewed-by: Kai Koehne <kai.koehne@qt.io>