Remove unique Android configuration from qtdeclarative.yml
The configuration is not used to test qtbase, so dependency updates
might fail if changes already in qtbase do not pass when then updating
dependencies of qtdeclarative.

This happens right now in
https://testresults.qt.io/coin/integration/qt/qtdeclarative/tasks/1677711115

If we want to enable this configuration anywhere we need to add it to
qtbase, or ideally the default list of platforms in cmake_platforms.yaml.

Pick-to: 6.4
Change-Id: I3f5614d34def41caf1747a5f8094bac3aa2d9e2b
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>