Update QDoc libclang to 15.0.0
The `libclang_version` variable is now updated to the current version, 15.0.0,
in both "coin/provisioning/common/unix/libclang.sh" and
"coin/provisioning/common/windows/libclang.ps1".

The naming of the artifacts for the provided library was slightly
changed from previous versions.

Before, the version number would contain two integers separated by a
dot.
The generated artifact would be named after this version number, with
the dot removed.

That is, for example, for version 12.0, a release artifact would be
called "libclang-release_120.*".

The updated two version uses a three segments version number and,
furthermore, the artifacts name do not collapse the separating dots such
that a release artifact would be called "libclang-release_15.0.0.*".

To support the new naming scheme, some processing was removed from both
provisioning scripts.

For Windows, the line removing all dots from the `libclang_version`
variable was removed.

For Unix, the removal of the dots in the `libclang_version`
variable when expanding the variable to generate the target urls to
retrieve the artifact was removed, using the bare value of the variable
itself.

For Unix provisioning, the artifacts for some of the platform, namely
"CentOS" and "Ubuntu 22.04", are not provided anymore.

"CentOS" was replaced by "Rhel8.4" on "gcc10.0", hence the target url
for the redhat platform was modified based on the new name.

The "Ubuntu 22.04" artifacts were not provided, such that the platform
was removed.

The sha1 of the target artifacts on both provisioning script was updated
for all platforms.

Task-number: QTBUG-107199
Pick-to: 6.4
Change-Id: I873680825b0953e897c610fb0b47f3cd90625646
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Kai Köhne <kai.koehne@qt.io>
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>