Fix ffmpeg build on windows CI
Ffmpeg was not compiled because of failed yasm installation
on msys2. The installation failed due to checking pgp signatures.
Current fix adds keys refreshing that fixes the problem (checked
on CI and on local machine)

Note, currently the ffmpeg build works with msvc and mingw,
additional settings for llvm should be added in next commits.

Build example:
https://testresults.qt.io/coin/integration/qt/qtmultimedia/tasks/web_qt_qtmultimedia_1667916985085

Pick-to: 6.4
Change-Id: Iff305e896bfb382c0c4030235cc2ad942cbbc002
Reviewed-by: Lars Knoll <lars@knoll.priv.no>