Work around glibc bug in our mdns resolution
glibc has a system where, if it fails to perform Happy Eyeballs host
name resolution in parallel, it will do it sequentially.
We triggered this because our docker setup was not responding to
systemd-resolved's AAAA queries over IPv6.
The bug in glibc was then that it would poll for 4999ms after sending
the first of the two queries. __Every time__ we do mdns host lookup.
In tst_QTcpSocket this happens multiple times in ::initTestCase for the
proxy global-datatags. That means we would be stuck polling for
4999ms*5 (number of proxy addresses to look up) before every test-case
that would use proxy (even if the test just does a return in the
proxy-case). Plus another 4999ms for any mdns lookup we would need to
perform during the test.
This led to the test severely slowing down and timing out quite often.

Work around this by doing full mDNS before trying DNS.
This will let us resolve multi-label mDNS addresses.

Pick-to: 6.5
Fixes: QTBUG-107696
Change-Id: I26c6b356b4c9712f4ac6d3613998ee448c0ca504
Reviewed-by: Dimitrios Apostolou <jimis@qt.io>