Coin/Android: Add Android 14 to CI
Add Android 14 to CI and keep 13 as maximum supported version, to allow
significant testing for both Android 13 and 14.

Android 13 is not yet removed, because Android 14 is still in beta
phase, so the plan is to fully replace Android 13 with Android 14 once
it's publicly released, and for now test Android 14, to avoid any
potential fails that might come before the final release.

[*] Android 14 is expected to be released late August 2023.

Task-number: QTQAINFRA-5609
Task-number: QTQAINFRA-5557
Change-Id: Ied5c560aa7d60f00948482fb1d7091b3139c6f0c
Reviewed-by: Simo Fält <simo.falt@qt.io>