COIN/ANDROID: Move android_emulator_launcher.sh to qt5.git
The android emulator launcher script is mainly used in CI and thus
makes sense to be moved there. This also makes it easier to make changes
without having to wait for submodule updates.

The env var ANDROID_EMULATOR_RUNNER points to the path of the runner
script, and can be used by qtbase.

qt5.git part of the change.

Task-number: QTQAINFRA-5596
Change-Id: If253e0a5f14d67526995561b1859312e61efdfff
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>