CMake: Ignore libraries provided by strawberry perl
Because we have strawberry perl's bin dir in the PATH, CMake considers
that a find_package prefix, and thus finds includes and libraries
that are shipped with strawberry perl.
That causes problems because the 3rd party runtime libraries are not
installed by the Qt installer, so any Qt modules / plugins that depend
on the 3rd party libraries will fail to load.

Use CMake 3.23's CMAKE_IGNORE_PREFIX_PATH option to tell it to ignore
strawberry headers and libraries.

We need to repeat it both for qtbase and non-qtbase args, to ensure
the libraries are not picked up for all of Qt.

This is a more general fix for problems like the one fixed by
a97bdde4d485fe0bde1a12c4307ae72a35b54c46

Task-number: QTBUG-113726
Change-Id: I706bf9ac945a47dc809bbb6eb9eda4ca93f84695
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>
(cherry picked from commit 9b7dcd94df9de3e98cd1831b7804fabfa97417dd)