Revert "Enable Unity Build on Qt Base for LLVM, and MinGW on Windows"
This reverts commit d1496ac752fc7cd4775a3fa45b704a775aa62fa9.

Unity builds have not been properly tested. In particular, they have
not been tested with an "infinite" batch-size (e.g. 100,000) to
reasonably ensure that any batch-size will work and any _new_ .cpp
files being added to SOURCES will break unity-build only in a way
that's fixable by changes in _that_ .cpp file.

The CI is not a lab for experiments. A unity-related failure has
brought the branch:dev CI to a screeching halt for a day on
2023-07-20, so we're disabling the use of it on the CI until it has
been properly tested in nightly builds and locally.

Task-number: QTBUG-109394
Change-Id: I935879b6b8a9602fcfbfcaa1d38b0b08254e0f59
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>