Revert "Enable Unity Build on Qt Base for macOS -developer-build"
This reverts commit 6f3ff63da5df242433e532a131736f8d1f964a78.

Unity builds have not been properly tested. In particular, they have
not been tested with an "infinite" batch-size (e.g. 100,000) to
reasonably ensure that any batch-size will work and any _new_ .cpp
files being added to SOURCES will break unity-build only in a way
that's fixable by changes in _that_ .cpp file.

The CI is not a lab for experiments. A unity-related failure has
brought the branch:dev CI to a screeching halt for a day on
2023-07-20, so we're disabling the use of it on the CI until it has
been properly tested in nightly builds and locally.

Task-number: QTBUG-109394
Change-Id: Ib1605c02bc088942acc7dfe90b3b0e39f46a41e3
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>