Revert "Enable Unity Build on QtBase for 3 Configurations"
This reverts commit 7b7dd85ca62dc53c21ea8272ff40aa60ce6288bf.

Unity builds have not been properly tested. In particular, they have
not been tested with an "infinite" batch-size (e.g. 100,000) to
reasonably ensure that any batch-size will work and any _new_ .cpp
files being added to SOURCES will break unity-build only in a way
that's fixable by changes in _that_ .cpp file.

The CI is not a lab for experiments. A unity-related failure has
brought the branch:dev CI to a screeching halt for a day on
2023-07-20, so we're disabling the use of it on the CI until it has
been properly tested in nightly builds and locally.

Task-number: QTBUG-109394
Change-Id: I0933d912215e8303e982c8649bc448103568ea65
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>