Android: Add missing ANDROID_EMULATOR param to Multi-ABI test target
The Multi-ABI targets run a subset of Android tests but the target
doesn't explicitly set the name of the Android emulator to run, so it
was defaulting to Android 6 because the CI Android emulator runner
script defaults to that. The latter is going change to be more explicit,
so this needs to be explicit also.

Pick-to: 6.6 6.5
Change-Id: Ifd5ddd5b5cf3b7a1b275f5a7b1252a21ae3165bc
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>