Doc: Create README template
There are many Qt submodules, and it would be beneficial to users if
there was a standard README template that will enable them to understand
and access the submodule content better. Currently many of the
submodules do not have README documents, or if they do, each submodule
varies vastly regarding what content is in the document. This proposed
README template is based on the Git ReadMe structure.

Task-number: QTBUG-114809
Pick-to: 6.5 6.6
Change-Id: I3d1d1da27ed665e76ee6d1df8a93c3e5a920b0a3
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>
Reviewed-by: Amir Masoud Abdol <amir.abdol@qt.io>