Add AAOS10 image for Android Automotive emulator as minimum supported SDK
Android Automotive should follow the lead for Q4A and regularly test
changes with both the minimum and then maximum supported SDK levels.
These are Android 10 and 13, of which 13 is already active on the CI.
This commit adds a Configuration for Android 10 and support for
creating an avd for that image.

Task-number: QAA-2073
Change-Id: I6d22c68f10a5da4b46664c67cf1369cbc7f5a537
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>