COIN/Android: Specify tags for emulator logcat output
Without specifying tags for logcat output, the flag -logcat-output does
not actually write anything to file. Looks like -logcat-output is
designed to take any logcat output that would be written to the
emulator output and write it to a file instead, so if there is no
logcat output in the emulator output, there will be no output to file.

Task-number: QTQAINFRA-5908
Change-Id: If44265c858f8cb6854659fed6dd15d02cbf1d7ab
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>