Stop testing macOS developer builds on x86_64 hardware
Our x86_64 pool is overloaded, so we need to reduce jobs.

As we're testing our packaging (universal) builds (both the one built
on x86_64 and the one built on arm64) on x86_64 for all macOS platforms
we support, it should be fine to remove the x86_64 tests of the
developer builds. These developer builds are still tested on arm64,
so the only test coverage we're losing on x86_64 are the subset of
tests that depend on QT_BUILD_INTERNAL.

Change-Id: Ia83964cc77d04ba796e3a59389ce000a51dcac93
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
(cherry picked from commit 760cf592664fa682962669e3449082a9a62d299d)
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>