Disable the truemotion1 decoder for the FFmpeg media backend
The truemotion1 decoder depends on a truemotion1data.h file which is
part of FFmpeg. This header file has a reference to GPL licensing, and
although the header provides a rationale for why the decoder tables are
not copyrightable, we prefer to not distribute this decoder as part of
the online Qt installer.

Users who need ability to decode truemotion1 format video files can do
so by recompiling the QtMultimedia FFmpeg media plugin against a custom
build of FFmpeg.

Task-number: QTBUG-118204
Pick-to: 6.6 6.5
Change-Id: Ia195ef9ce576d18047e7b38bbc5019648ba42b4c
Reviewed-by: Kai Köhne <kai.koehne@qt.io>
(cherry picked from commit 027acd91fd68ef748b81842ecefa8d471a480bd8)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>